fix: itinerary list shows when page is loaded under the typewriter effect
