+++ package.json
@@ -22,6 +22,7 @@
     "react": "^18",
     "react-dom": "^18",
     "react-icons": "^5.3.0",
+    "react-image-gallery": "^1.3.0",
     "react-select": "^5.8.0",
     "react-slider": "^2.0.6",
     "react-tag-input": "^6.10.3",
