+++ src/app/utils/utils.js
@@ -23,12 +23,28 @@ function generateBudgetString(length, char) {
   return Array.from({ length }, () => char).join("");
 }
 
-function filterItineraries(itineraries, locationArr, budget) {
+function filterItineraries(
+  allItineraries,
+  locationArr,
+  budget,
+  minDays,
+  maxDays
+) {
+  minDays ? null : (minDays = 0);
+  maxDays ? null : (maxDays = 99999999);
+
   let filteredMatchScores = [];
+  let itineraries = allItineraries.filter((itinerary) => {
+    return (
+      itinerary.number_of_days >= minDays && itinerary.number_of_days <= maxDays
+    );
+  });
+
   itineraries.forEach((itinerary) => {
     itinerary.matchScore = 0;
     filteredMatchScores.push(itinerary);
   });
+
   if (locationArr.length > 0) {
     const matchScores = itineraries.map((itinerary) => {
       locationArr.forEach((location) => {
